fix silu

+++ ivy/functional/backends/numpy/activations.py
@@ -41,7 +41,7 @@ def gelu(x, /, *, approximate: Optional[bool] = True) -> np.ndarray:
 
 
 def sigmoid(x: np.ndarray, /) -> np.ndarray:
-    return np.asarray(1 / (1 + np.exp(-x))).astype(x.dtype)
+    return np.asarray(1 / (1 + np.exp(-x)))
 
 
 def softmax(
